requirements.txt
amqp==2.1.1
anyjson==0.3.3
Babel==2.3.4
beautifulsoup4==4.5.1
billiard==3.5.0.2
blinker==1.4
bs4==0.0.1
celery==4.0.0
click==6.6
coverage==4.2
decorator==4.0.10
defusedxml==0.4.1
elasticsearch==5.0.1
elasticsearch-dsl==5.0.0
Flask==0.11.1
Flask-Login==0.4.0
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.1
Flask-WTF==0.13.1
flipflop==1.0
guess-language==0.2
itsdangerous==0.24
Jinja2==2.8
kombu==4.0.0
MarkupSafe==0.23
pbr==1.10.0
psycopg2==2.6.2
pycurl==7.43.0
PySocks==1.5.7
python-dateutil==2.6.0
pytz==2016.7
requests==2.11.1
selenium==3.0.1
six==1.10.0
SQLAlchemy==1.1.3
sqlalchemy-migrate==0.10.0
sqlparse==0.2.2
stem==1.4.0
Tempita==0.5.2
urllib3==1.19
vine==1.1.3
Werkzeug==0.11.11
WTForms==2.1
